requirements.txt
mysql-connector-python
tkcalendar
openpyxl
tkinter